Move IP util to `util` module
