+++ src/ui/widgets/main_window.py
@@ -1,6 +1,6 @@
 
 from PyQt5.QtWidgets import *
-from .getIP import get_ip_thru_gateway as get_ip
+from util.net import get_ip_thru_gateway as get_ip
 
 def create_main_window(title: str, callbacks) -> QWidget:
     '''
